+++ bindings/vue/CHANGELOG.md
@@ -2,6 +2,12 @@ CHANGELOG
 ====
 
 dev
+----
+
+### New features
+
+* VOnsCard: New component.
+
 2.0.0-beta.4
 ----